Remove vim dependency from graphlog, move functions to util library.
