+++ autoload/mundo/util.py
@@ -1,5 +1,8 @@
 import vim
 
+normal = lambda s: vim.command('normal %s' % s)
+normal_silent = lambda s: vim.command('silent! normal %s' % s)
+
 def _goto_window_for_buffer(b):
     w = int(vim.eval('bufwinnr(%d)' % int(b)))
     vim.command('%dwincmd w' % int(w))
@@ -7,6 +10,13 @@ def _goto_window_for_buffer(b):
 def _goto_window_for_buffer_name(bn):
     b = vim.eval('bufnr("%s")' % bn)
     return _goto_window_for_buffer(b)
+
+# Rendering utility functions
+def _output_preview_text(lines):
+    _goto_window_for_buffer_name('__Gundo_Preview__')
+    vim.command('setlocal modifiable')
+    vim.current.buffer[:] = [line.rstrip() for line in lines]
+    vim.command('setlocal nomodifiable')
 
 def _undo_to(n):
     n = int(n)
